% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate Quarto HTML / PDF Country Factsheets in Batch}
\usage{
generate_report(type = "donor", name = NULL)
}
\arguments{
\item{type}{either donor recipient or destination.}

\item{name}{name of  donor recipient or destination if NULL batch process them}
}
\value{
nothing, reports are generated in the specified folder.
}
\description{
This function renders the 'Country_Factsheet' Quarto template for a list of
countries, generating a report file for each.
}
\examples{
#generate_report(type = "donor")
}
