% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_destination_funding_trend.R
\name{indicator_destination_funding_trend}
\alias{indicator_destination_funding_trend}
\title{Destination-Focused: Funding Growth Trend}
\usage{
indicator_destination_funding_trend(flows, destinations_df = NULL)
}
\arguments{
\item{flows}{A dataframe with \code{destinationObjects}, \code{budgetYear},
and \code{amountUSD}.}

\item{destinations_df}{(optional) A dataframe of destinations to merge
results with.}
}
\value{
A tibble with \code{destination} and \code{Destination_Funding_Growth_Trend}.
}
\description{
Computes the slope of the linear regression of funding over time
for each destination.
Goal: Detects whether funding to a destination is increasing or decreasing
over time.
}
\examples{
destination <- indicator_destination_funding_trend(flows)
#table(destination$Destination_Funding_Growth_Trend)

library(ggplot2)
ggplot(destination, aes(x = Destination_Funding_Growth_Trend)) +
  geom_histogram( fill = "red", color = "white") +
  unhcrthemes::theme_unhcr(grid = "Y", axis = "X", axis_title = FALSE) +
  labs(
    title = "Distribution of Destination Funding Growth Trend",
    x = "",
    y = "Number of Destination",
    caption = paste(
    "Indicator interpretation:",
    "Linear regression slope of annual funding over time. 
    Positive values=increasing funding, negative values=decreasing funding,
    zero=stable funding.", "\n\n",
    
    "How to read the visualization:",
    "Histogram shows growth/decline patterns. Right of zero indicates expanding
    crisis responses, left indicates potential donor fatigue or improving 
    contexts.", "\n\n",
    
    "Humanitarian relevance:",
    "Identifies crises receiving increasing attention vs. those facing funding
    declines despite ongoing needs, informing advocacy priorities.", "\n\n",
    
    "Source: Financial Tracking Service (FTS) API."
  )
  )
}
