% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_donor_recipient_concentration.R
\name{indicator_donor_recipient_concentration}
\alias{indicator_donor_recipient_concentration}
\title{Donor→Recipient: Funding Concentration Index}
\usage{
indicator_donor_recipient_concentration(
  flows,
  donor_recipient_pairs = NULL,
  donors = NULL,
  recipient_name = NULL
)
}
\arguments{
\item{flows}{A dataframe with \code{sourceObjects}, \code{destinationObjects},
and \code{amountUSD}.}

\item{donor_recipient_pairs}{(optional) A dataframe of donor–recipient pairs to merge
results with.}

\item{donors}{(optional used if recipient_name is provided ) A dataframe with donor names to merge with the  results.}

\item{recipient_name}{(optional) the name of a specific recipient to filter the results of the function}
}
\value{
A tibble with \code{donor}, \code{recipient}, and
\code{DonorRecipient_Concentration_Index}.
}
\description{
Measures the share of a recipient’s total funding that comes from a
specific donor.
}
\details{
Goal: Quantifies how dependent the recipient is on that donor (share of
donor’s contribution among total received).
}
\examples{
donor_recipient <- indicator_donor_recipient_concentration(flows)
#table(donor_recipient$DonorRecipient_Stability_Index)

library(ggplot2)
ggplot(donor_recipient, aes(x = DonorRecipient_Concentration_Index)) +
  geom_histogram( fill = "green", color = "white") +
  unhcrthemes::theme_unhcr(grid = "Y", axis = "X", axis_title = FALSE) +
  labs(
    title = "Distribution of Donor Recipient Concentration Index",
    x = "",
    y = "Number of Donor-Recipient Pairs",
    caption = paste(
    "Indicator interpretation:",
    "Share of recipient's total funding coming from specific donor (0-1). 
    Higher values indicate recipient dependency on that donor.", "\n\n",
    
    "How to read the visualization:",
    "Histogram shows dependency levels across partnerships. Values near 1 
    indicate critical donor dependencies requiring diversification strategies.",
    "\n\n",
    
    "Humanitarian relevance:",
    "Identifies vulnerable partnerships where recipient organizations risk 
    severe disruption if key donors change funding priorities.", "\n\n",
    
    "Source: Financial Tracking Service (FTS) API."
  )
  )
}
