% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indicator_donor_recipient_relationship_duration.R
\name{indicator_donor_recipient_relationship_duration}
\alias{indicator_donor_recipient_relationship_duration}
\title{Donor-Recipient Relationship Duration Index}
\usage{
indicator_donor_recipient_relationship_duration(
  flows,
  donor_recipient_pairs = NULL,
  donors = NULL,
  recipient_name = NULL
)
}
\arguments{
\item{flows}{A dataframe containing \code{sourceObjects}, \code{destinationObjects}, and \code{budgetYear}.}

\item{donor_recipient_pairs}{(optional) A dataframe of donor–recipient pairs to merge
results with.}

\item{donors}{(optional used if recipient_name is provided ) A dataframe with donor names to merge with the  results.}

\item{recipient_name}{(optional) the name of a specific recipient to filter the results of the function}
}
\value{
A tibble with columns \code{donor}, \code{recipient}, and \code{Relationship_Duration_Index}.
}
\description{
Calculates the normalized duration of active funding between a donor and recipient.
}
\examples{
donor_recipient <- indicator_donor_recipient_relationship_duration(flows)
#table(donor_recipient$Relationship_Duration_Index)

library(ggplot2)
ggplot(donor_recipient, aes(x = Relationship_Duration_Index)) +
  geom_histogram( fill = "green", color = "white") +
  unhcrthemes::theme_unhcr(grid = "Y", axis = "X", axis_title = FALSE) +
  labs(
    title = "Distribution of Donor Recipient Relationship Duration Index",
    x = "",
    y = "Number of Donor-Recipient Pairs",
    caption = paste(
    "Indicator interpretation:",
    "Normalized duration of active funding relationships between donor-recipient
    pairs (0=new relationship, 1=active since earliest data).", "\n\n",
    
    "How to read the visualization:",
    "Histogram shows relationship longevity across partnerships. Right-skewed 
    distribution indicates many new relationships, left-skewed indicates 
    established partnerships.", "\n\n",
    
    "Humanitarian relevance:",
    "Long-term partnerships build trust, contextual understanding, and
    operational efficiency, leading to more effective crisis response.", "\n\n",
    
    "Source: Financial Tracking Service (FTS) API."
  )
  )
}
