% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_donor_relationship_duration.R
\name{indicator_donor_relationship_duration}
\alias{indicator_donor_relationship_duration}
\title{Relationship Duration Index}
\usage{
indicator_donor_relationship_duration(flows, donors = NULL)
}
\arguments{
\item{flows}{A dataframe containing \code{sourceObjects} and \code{budgetYear}.}

\item{donors}{Optional dataframe with donor names.}
}
\value{
A tibble with \code{donor} and \code{Relationship_Duration_Index}.
}
\description{
Measures the duration of a donor’s funding relationship relative to the
maximum possible duration.
}
\examples{
donors <- indicator_donor_relationship_duration(flows)
#table(donor$Relationship_Duration_Index)
library(ggplot2)
ggplot(donors, aes(x = Relationship_Duration_Index)) +
  geom_histogram(binwidth = 0.05, fill = "steelblue", color = "white") +
  unhcrthemes::theme_unhcr(grid = "Y", axis = "X", axis_title = FALSE) +
  labs(
    title = "Distribution of Donor Relationship Duration Index",
    x = "",
    y = "Number of Donors",
    caption = paste(
    "Indicator interpretation:",
    "Measures duration of donor's funding relationship relative to maximum 
    possible duration. 1 = active since earliest data, 0 = new entrant.", "\n\n",
    
    "How to read the visualization:",
    "Histogram shows distribution of duration scores. Left-skewed distribution 
    indicates many new donors, right-skewed indicates established 
    relationships.", "\n\n",
    
    "Humanitarian relevance:",
    "Long-term donors often have deeper contextual understanding and more 
    predictable funding patterns for strategic planning.", "\n\n",
    
    "Source: Financial Tracking Service (FTS) API."
  )
  )
}
