% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_destination_funding_destination.R
\name{plot_destination_funding_destination}
\alias{plot_destination_funding_destination}
\title{Plot Funding by Destination Location}
\usage{
plot_destination_funding_destination(flows, destination_name)
}
\arguments{
\item{flows}{A dataframe of humanitarian funding flows (FTS format).}

\item{destination_name}{The name of the destination location to filter.}
}
\value{
A ggplot object showing total funding per donor for the destination.
}
\description{
Visualizes total humanitarian funding directed to a given destination destination.
Useful to identify funding hotspots and compare regions.
}
\examples{
p <- plot_destination_funding_destination(flows, destination_name= "Burundi")

# getting LLm story
story <- generate_plot_story(p, provider = "azure", model = "gpt-4.1-mini", max_tokens = 300)
# cat(story)
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p, provider = "azure", model = "gpt-4.1-mini")
p + ggplot2::labs(subtitle = dubbed)
}
