% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_donor_earmarking.R
\name{plot_donor_earmarking}
\alias{plot_donor_earmarking}
\title{Plot Donor Funding by Earmarking Type}
\usage{
plot_donor_earmarking(flows, donor_name)
}
\arguments{
\item{flows}{A dataframe of humanitarian funding flows (FTS format).}

\item{donor_name}{Donor organization name to filter.}
}
\value{
A ggplot object showing donor funding by earmarking type.
}
\description{
Visualizes the proportion of funding per earmarking type
(Unearmarked, Softly earmarked, etc.)
for a selected donor. Helps understand donor flexibility and restrictions.
}
\examples{
p <- plot_donor_earmarking(flows, donor_name= "Spain, Government of")
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p, provider = "azure", model = "gpt-4.1-mini")
p + ggplot2::labs(subtitle = dubbed)
}
