% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_location_funding_destination.R
\name{plot_location_funding_destination}
\alias{plot_location_funding_destination}
\title{Plot Funding by Destination Location}
\usage{
plot_location_funding_destination(flows, location_name)
}
\arguments{
\item{flows}{A dataframe of humanitarian funding flows (FTS format).}

\item{location_name}{The name of the destination location to filter.}
}
\value{
A ggplot object showing total funding per donor for the location.
}
\description{
Visualizes total humanitarian funding directed to a given destination location.
Useful to identify funding hotspots and compare regions.
}
\examples{
p <- plot_location_funding_destination(flows, location_name= "Burundi")

# getting LLm story
story <- generate_plot_story(p, max_tokens = 300)
cat(story)
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p)
p + ggplot2::labs(subtitle = dubbed)
}
