% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_location_funding_donor.R
\name{plot_location_funding_donor}
\alias{plot_location_funding_donor}
\title{Plot Destination Funding by Donor}
\usage{
plot_location_funding_donor(flows, location_name)
}
\arguments{
\item{flows}{FTS flows dataframe.}

\item{location_name}{Destination location name.}
}
\value{
A ggplot object showing total funding per donor for the location.
}
\description{
Shows funding received at a location from multiple donors.
Useful to identify funding hotspots or dependency.
}
\examples{
p <- plot_location_funding_donor(flows, location_name= "Burundi")

# getting LLm story
story <- generate_plot_story(p, max_tokens = 300)
cat(story)
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p)
p + ggplot2::labs(subtitle = dubbed)
}
