% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_location_heatmap.R
\name{plot_location_heatmap}
\alias{plot_location_heatmap}
\title{Plot Top Donors Heatmap}
\usage{
plot_location_heatmap(flows, location_name, top_n = 5)
}
\arguments{
\item{flows}{FTS flows dataframe}

\item{location_name}{Name of the destination location to filter for}

\item{top_n}{Number of top donors to display (default = 5)}
}
\value{
A ggplot heatmap
}
\description{
Visualizes the total funding from top donors to recipient organizations
for a given location. Correctly handles deeply nested FTS destinationObjects.
}
\examples{
p <- plot_location_heatmap(flows, location_name= "Burundi", top_n = 5)

# getting LLm story
story <- generate_plot_story(p, max_tokens = 300)
cat(story)
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p)
p + ggplot2::labs(subtitle = dubbed)
}
