% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recipient_contribution_type.R
\name{plot_recipient_contribution_type}
\alias{plot_recipient_contribution_type}
\title{Plot Recipient Funding by Contribution Type}
\usage{
plot_recipient_contribution_type(flows, recipient_name)
}
\arguments{
\item{flows}{FTS flows dataframe.}

\item{recipient_name}{Recipient organization name.}
}
\value{
A ggplot object showing funding by contribution type.
}
\description{
Visualizes the split of funding received by a recipient based on
contribution type
(e.g., financial or in-kind). Useful to assess funding modality diversity.
}
\examples{
p <- plot_recipient_contribution_type(flows, 
                recipient_name = "United Nations High Commissioner for Refugees")


# getting LLm story
story <- generate_plot_story(p, provider = "azure", model = "gpt-4.1-mini", max_tokens = 300)
# cat(story)
# and plot with more powerful subtitle
dubbed <- generate_plot_story(p, provider = "azure", model = "gpt-4.1-mini")
p + ggplot2::labs(subtitle = dubbed)
}
